show connectionLoss/COMM_ALARM when monitoredItem setup failed
Value would appear frozen when installing monitoredItem failed for
example because the server has reached the limit of monitored items
or if an item cannot be monitored because it used invalid filters
like a deadband filter on a string.

show connectionLoss/COMM_ALARM when monitoredItem setup failed

Value would simply appear frozen otherwise.
Needed to revert order of initial read and montior setup so that
errors in monitor setup are not cleared by successful initial read.