Don't process writeComplete on struct elements that have not been written
If writing to an optstruct or union, processing the records for all elements,
not only the one that has been written to, with a writeComplete callback
would clear errors for non-existing options and not selected choices.
This clears the INVALID alarm for those elements.
This becomes visible only if the value did not change and thus the optsruct
or union did not send back a monitor update.